готовый

+++ app/Http/Controllers/AllController.php
@@ -49,14 +49,29 @@ class AllController extends Controller
     public function selected()
     {
         $selecteds = Selected::get();
-        return view('ready.selected', compact('selecteds'));
+        return view('ready.selected-section', compact('selecteds'));
     }
 
     public function selectedPost(Request $request)
     {
         if( $request->ajax() ) {
-            $studId = $request->studid;
-            Selected::where('id', $studId)->delete();
+            if( $request->studid ) {
+                $studId = $request->studid;
+                Selected::where('id', $studId)->delete();
+            }
+            elseif( $request->data || $request->workerireset ) {
+                $workerid = $request->workerireset;
+                $surname = $request->data[0]['surname'];
+                $name = $request->data[0]['name'];
+                $lastname = $request->data[0]['lastname'];
+                $position = $request->data[0]['position'];
+                DB::table('selecteds')->where('id', $workerid)->update([
+                    'surname' => trim($surname),
+                    'name' => trim($name),
+                    'lastname' => trim($lastname),
+                    'position' => trim($position)
+                ]);
+            }
         }
     }
 
@@ -69,29 +84,35 @@ class AllController extends Controller
     public function workersAdd(Request $request)
     {
         if( $request->ajax() ) {
-            $data = $request->workerid;
-            
-            $StudName = Worker::where('id', $data)->value('name');
-            $StudSurname = Worker::where('id', $data)->value('surname');
-            $StudLastname = Worker::where('id', $data)->value('lastname');
-            $StudPosition = Worker::where('id', $data)->value('position');
-            $issetName = Selected::where([
-                'name' => $StudName,
-                'surname' => $StudSurname,
-                'lastname' => $StudLastname,
-                'position' => $StudPosition
-            ])->value('id');
-            // dd(1);
-            if( $issetName ) {
-                return 1;
-            }else {
-                DB::table('selecteds')->insert([
+            if( $request->workeriddelete ) {
+                $workeriddelete = $request->workeriddelete;
+                Worker::where('id', $workeriddelete)->delete();
+            }
+            elseif( $request->workerid ){
+                $data = $request->workerid;
+                
+                $StudName = Worker::where('id', $data)->value('name');
+                $StudSurname = Worker::where('id', $data)->value('surname');
+                $StudLastname = Worker::where('id', $data)->value('lastname');
+                $StudPosition = Worker::where('id', $data)->value('position');
+                $issetName = Selected::where([
                     'name' => $StudName,
                     'surname' => $StudSurname,
                     'lastname' => $StudLastname,
                     'position' => $StudPosition
-                ]);
-                // return redirect()->back();
+                ])->value('id');
+                // dd(1);
+                if( $issetName ) {
+                    return 1;
+                }else {
+                    DB::table('selecteds')->insert([
+                        'name' => $StudName,
+                        'surname' => $StudSurname,
+                        'lastname' => $StudLastname,
+                        'position' => $StudPosition
+                    ]);
+                    // return redirect()->back();
+                }
             }
         }
     }
